# docs/api/main-backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "RoomPACT Main Backend",
    "description": "Main backend API for RoomPACT Campus",
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "/api",
      "description": "Public API behind nginx on EC2"
    },
    {
      "url": "http://localhost:8000",
      "description": "Direct local development"
    }
  ],
  "paths": {
...
    "/profiles": {
      "post": {
        "tags": [
          "profiles"
        ],
        "summary": "Create Profile",
        "operationId": "create_profile_profiles_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Profile Profiles Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "profiles"
        ],
        "summary": "List Profiles",
        "operationId": "list_profiles_profiles_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Profiles Profiles Get"
                }
              }
            }
          }
        }
      }
    },
...
    "/profiles/{profile_id}/match-requests": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Match Request",
        "operationId": "create_match_request_profiles__profile_id__match_requests_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRoomCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Match Request Profiles  Profile Id  Match Requests Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "List Match Requests",
        "operationId": "list_match_requests_profiles__profile_id__match_requests_get",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response List Match Requests Profiles  Profile Id  Match Requests Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profiles/{profile_id}/chat-rooms": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Create Chat Room",
        "operationId": "create_chat_room_profiles__profile_id__chat_rooms_post",
        "parameters": [
          {
            "name": "profile_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Profile Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChatRoomCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Create Chat Room Profiles  Profile Id  Chat Rooms Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/match-requests/{request_id}/accept": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Accept Match Request",
        "operationId": "accept_match_request_match_requests__request_id__accept_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequestAcceptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Accept Match Request Match Requests  Request Id  Accept Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/match-requests/{request_id}/reject": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Reject Match Request",
        "operationId": "reject_match_request_match_requests__request_id__reject_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchRequestAcceptRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Reject Match Request Match Requests  Request Id  Reject Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat-rooms/{room_id}/messages": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat Room Messages",
        "operationId": "get_chat_room_messages_chat_rooms__room_id__messages_get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Chat Room Messages Chat Rooms  Room Id  Messages Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat-rooms/{room_id}/roommate-confirmation": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Confirm Roommate",
        "operationId": "confirm_roommate_chat_rooms__room_id__roommate_confirmation_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoommateConfirmationRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Confirm Roommate Chat Rooms  Room Id  Roommate Confirmation Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat-rooms/{room_id}/pact": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Roommate Pact",
        "operationId": "get_roommate_pact_chat_rooms__room_id__pact_get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Roommate Pact Chat Rooms  Room Id  Pact Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "chat"
        ],
        "summary": "Update Roommate Pact",
        "operationId": "update_roommate_pact_chat_rooms__room_id__pact_put",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PactUpdateRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Update Roommate Pact Chat Rooms  Room Id  Pact Put"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat-rooms/{room_id}/question-suggestions": {
      "get": {
        "tags": [
          "chat"
        ],
        "summary": "Get Chat Question Suggestions",
        "operationId": "get_chat_question_suggestions_chat_rooms__room_id__question_suggestions_get",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          },
          {
            "name": "profile_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "title": "Profile Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Get Chat Question Suggestions Chat Rooms  Room Id  Question Suggestions Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/chat-rooms/{room_id}/signatures": {
      "post": {
        "tags": [
          "chat"
        ],
        "summary": "Sign Roommate Pact",
        "operationId": "sign_roommate_pact_chat_rooms__room_id__signatures_post",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Room Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PactSignatureRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Sign Roommate Pact Chat Rooms  Room Id  Signatures Post"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ChatRoomCreateRequest": {
        "properties": {
          "other_profile_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Other Profile Id"
          }
        },
        "type": "object",
        "required": [
          "other_profile_id"
        ],
        "title": "ChatRoomCreateRequest"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MatchRequestAcceptRequest": {
        "properties": {
          "profile_id": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Profile Id"
          }
        },
        "type": "object",
        "required": [
          "profile_id"
        ],
        "title": "MatchRequestAcceptRequest"
      },
...
      "ProfileCreateRequest": {
        "properties": {
          "nickname": {
            "type": "string",
            "maxLength": 30,
            "minLength": 1,
            "title": "Nickname"
          },
          "age": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 17.0,
            "title": "Age"
          },
          "gender": {
            "type": "string",
            "maxLength": 20,
            "minLength": 1,
            "title": "Gender"
          },
          "region": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Region"
          },
          "move_in_period": {
            "type": "string",
            "maxLength": 50,
            "minLength": 1,
            "title": "Move In Period"
          },
          "stay_duration_months": {
            "type": "integer",
            "maximum": 120.0,
            "minimum": 1.0,
            "title": "Stay Duration Months"
          }
        },
        "type": "object",
        "required": [
          "nickname",
          "age",
          "gender",
          "region",
          "move_in_period",
          "stay_duration_months"
        ],
        "title": "ProfileCreateRequest"
      },
...
    }
  }
}